backport of commit c4621edec7c10963740a266bd145f250c67cafaa
