+++ ui/app/templates/workspace/projects/project/app/deployments.hbs
@@ -1,5 +1,5 @@
 {{page-title (concat @model.application.application "Deployments")}}
-{{#let (not-eq this.target.currentRouteName 'workspace.projects.project.app.deployment.deployment-seq.resource') as |isNotResourcePage|}}
+{{#let (not-eq this.target.currentRouteName 'workspace.projects.project.app.deployments.deployment.resource') as |isNotResourcePage|}}
   <div class={{if (and isNotResourcePage this.deploymentsByGeneration) "grid grid--menu-with-panel"}}>
   {{#if isNotResourcePage}}
     {{#if this.deploymentsByGeneration}}
