+++ ui/app/templates/workspace/projects/project/app/deployments/deployment/resource.hbs
