+++ ui/tests/acceptance/deployment-detail-test.ts
@@ -1,4 +1,4 @@
-import { currentURL, visit } from '@ember/test-helpers';
+import { visit } from '@ember/test-helpers';
 import { module, test } from 'qunit';
 
 import { setupApplicationTest } from 'ember-qunit';
@@ -16,7 +16,7 @@ module('Acceptance | deployment detail', function (hooks) {
     let deployment = this.server.create('deployment', 'random', { application });
     this.server.create('status-report', 'ready', { application, target: deployment });
 
-    await visit(`/default/acme-project/app/acme-app/deployment/seq/${deployment.sequence}`);
+    await visit(`/default/acme-project/app/acme-app/deployments/${deployment.sequence}`);
 
     assert.dom('[data-test-status-report-indicator="ready"]').exists();
   });
@@ -26,18 +26,8 @@ module('Acceptance | deployment detail', function (hooks) {
     let application = this.server.create('application', { name: 'acme-app', project });
     let deployment = this.server.create('deployment', 'random', { application });
 
-    await visit(`/default/acme-project/app/acme-app/deployment/seq/${deployment.sequence}`);
+    await visit(`/default/acme-project/app/acme-app/deployments/${deployment.sequence}`);
 
     assert.dom('[data-test-status-report-indicator]').doesNotExist();
-  });
-
-  test('redirects from the ID to the sequence route', async function (assert) {
-    let project = this.server.create('project', { name: 'acme-project' });
-    let application = this.server.create('application', { name: 'acme-app', project });
-    let deployment = this.server.create('deployment', 'random', { application });
-
-    await visit(`/default/acme-project/app/acme-app/deployment/${deployment.id}`);
-
-    assert.equal(currentURL(), `/default/acme-project/app/acme-app/deployment/seq/${deployment.sequence}`);
   });
 });
